Refactor PerpetualConstruction to its own file
